+++ client/package.json
@@ -7,6 +7,8 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
+    "@types/react": "^18.2.45",
+    "@types/react-dom": "^18.2.18",
     "cors": "^2.8.5",
     "date-fns": "^2.30.0",
     "express": "^4.18.2",
