+++ client/src/components/Home/Body.tsx
@@ -1,4 +1,4 @@
-import { React, Fragment, useEffect, useRef, useState } from "react";
+import React from "react";
 
 function Body(props) {
   return (
