--- client/src/components/Home/Footer.js
@@ -1,12 +1,0 @@
-import { React, Fragment, useEffect, useRef, useState } from "react";
-
-function Footer(props) {
-  return (
-    <div>
-      <li>https://www.linkedin.com/in/seany4ng/</li>
-      <li>https://www.linkedin.com/in/ye-oliver/</li>
-    </div>
-  );
-}
-
-export default Footer;
