+++ client/src/components/Home/Footer.tsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Footer(props) {
+  return (
+    <div>
+      footer component
+    </div>
+  );
+}
+
+export default Footer;
