+++ client/src/components/Home/Header.tsx
@@ -1,4 +1,4 @@
-import { React, Fragment, useEffect, useRef, useState } from "react";
+import React from "react";
 
 function Header(props) {
   return <div>header component</div>;
