+++ client/src/pages/Profile.tsx
@@ -1,4 +1,4 @@
-import { React, Fragment, useEffect, useRef, useState } from "react";
+import React from "react";
 
 function Profile(props) {
   return (
